suse-commit: 7069df54bcbf97d63d9101d281e8ceb3076af84d

+++ drivers/xen/scsiback/common.h
@@ -61,7 +61,7 @@ struct ids_tuple {
 	unsigned int hst;		/* host    */
 	unsigned int chn;		/* channel */
 	unsigned int tgt;		/* target  */
-	unsigned int lun;		/* LUN     */
+	u64 lun;			/* LUN     */
 };
 
 struct v2p_entry {
